+++ jest.config.js
@@ -21,7 +21,7 @@ module.exports = {
     moduleFileExtensions: ['js', 'json', 'jsx'],
   
     // The paths to modules that run some code to configure or set up the testing environment before each test
-    setupFiles: ['<rootDir>/tests/setup.js'],
+    setupFiles: ['<rootDir>/src/tests/setup.js'],
   
     // The test environment that will be used for testing
     testEnvironment: 'jsdom',
