+++ src/tests/adminTest/addParties.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import expect from 'expect';
 import { shallow } from 'enzyme';
-import {  AddParty } from '../../src/components/pages/adminPages/add-parties.jsx';
+import {  AddParty } from '../../components/pages/adminPages/add-parties.jsx';
 
 describe('AddParties component should', () => {
     const props = {
