+++ src/tests/adminTest/parties.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import expect from 'expect';
 import { shallow } from 'enzyme';
-import { Parties } from '../../src/components/pages/adminPages/parties.jsx';
+import { Parties } from '../../components/pages/adminPages/parties.jsx';
 
 describe('Parties component should', () => {
     const props = {
