+++ src/tests/adminTest/register.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import expect from 'expect';
 import { shallow } from 'enzyme';
-import { Register } from '../../src/components/pages/adminPages/register-candidate.jsx';
+import { Register } from '../../components/pages/adminPages/register-candidate.jsx';
 
 describe('Register component should', () => {
     const props = {
