+++ src/tests/electionResult.test.js
@@ -3,44 +3,50 @@ import { shallow, mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import AllOfficeStore, { Alloffice } from '../src/components/pages/govOffice-page.jsx';
-import UserNavbar from '../src/components/reuseable component/user-navbar.component.jsx';
+import ElectionResultStore, { Results } from '../components/pages/election-results.jsx';
+import UserNavbar from '../components/reuseable component/user-navbar.component.jsx';
 
 
 const props = {
-  items: {
-},
-  getAllOffice: jest.fn(() => Promise.resolve())
+  posts: {},
+  getAllElectionResults: jest.fn(() => Promise.resolve()),
+  candidates: [{
+    candidate:'',
+    result:'',
+}],
+match: {
+  params: {
+    officeId:1
+  }
+}
 };
 
 const initialState = {
   get: {
       success: true,
-      items:{
-      data: {
+      candidates: {
           title: '',
           content: ''
       },
-    },
       error: ''
   },
-}
+};
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 const store = mockStore(initialState);
 
-const instance = () => mount(<Alloffice />);
+const instance = () => mount(<ElectionResultStore />);
 
 const OfficeComponent = (
   <Provider store={store}>
-    <Alloffice {...props} />
+    <Results {...props} />
   </Provider>
 );
 
 describe('Alloffice component', () => {
      it('render component', () => {
-      const wrapper = shallow(<Alloffice {...props} />);
+      const wrapper = shallow(<Results {...props} />);
       expect(wrapper).toMatchSnapshot();
      });
 });
@@ -57,7 +63,7 @@ describe('Office component', () => {
   });
 
   it('has map state to props', () => {
-    const component = shallow(<AllOfficeStore {...props} store={store} />);
+    const component = shallow(<ElectionResultStore {...props} store={store} />);
     expect(component.length).toBe(1);
   });
