+++ src/tests/govOffice.test.js
@@ -3,50 +3,44 @@ import { shallow, mount } from 'enzyme';
 import { Provider } from 'react-redux';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import ElectionResultStore, { Results } from '../src/components/pages/election-results.jsx';
-import UserNavbar from '../src/components/reuseable component/user-navbar.component.jsx';
+import AllOfficeStore, { Alloffice } from '../components/pages/govOffice-page.jsx';
+import UserNavbar from '../components/reuseable component/user-navbar.component.jsx';
 
 
 const props = {
-  posts: {},
-  getAllElectionResults: jest.fn(() => Promise.resolve()),
-  candidates: [{
-    candidate:'',
-    result:'',
-}],
-match: {
-  params: {
-    officeId:1
-  }
-}
+  items: {
+},
+  getAllOffice: jest.fn(() => Promise.resolve())
 };
 
 const initialState = {
   get: {
       success: true,
-      candidates: {
+      items:{
+      data: {
           title: '',
           content: ''
       },
+    },
       error: ''
   },
-};
+}
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 const store = mockStore(initialState);
 
-const instance = () => mount(<ElectionResultStore />);
+const instance = () => mount(<Alloffice />);
 
 const OfficeComponent = (
   <Provider store={store}>
-    <Results {...props} />
+    <Alloffice {...props} />
   </Provider>
 );
 
 describe('Alloffice component', () => {
      it('render component', () => {
-      const wrapper = shallow(<Results {...props} />);
+      const wrapper = shallow(<Alloffice {...props} />);
       expect(wrapper).toMatchSnapshot();
      });
 });
@@ -63,7 +57,7 @@ describe('Office component', () => {
   });
 
   it('has map state to props', () => {
-    const component = shallow(<ElectionResultStore {...props} store={store} />);
+    const component = shallow(<AllOfficeStore {...props} store={store} />);
     expect(component.length).toBe(1);
   });
