+++ src/tests/homepage.test.js
@@ -3,9 +3,9 @@ import { shallow } from 'enzyme';
 import { Provider } from 'react-redux';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import HomePage from '../src/components/pages/home-page.jsx';
-import Navbar from '../src/components/reuseable component/navbar-component.jsx';
-import Footer from '../src/components/reuseable component/footer-component.jsx';
+import HomePage from '../components/pages/home-page.jsx';
+import Navbar from '../components/reuseable component/navbar-component.jsx';
+import Footer from '../components/reuseable component/footer-component.jsx';
 
 const middleware = [thunk];
 const mockStore = configureMockStore(middleware);
@@ -19,7 +19,6 @@ const HomePageComponent = (
 
 describe('HomePage component', () => {
   it('should match snapshot', () => {
-    // const wrap = HomePageComponent;
     const wrapper = shallow(<HomePage />);
     expect(wrapper).toMatchSnapshot();
   });
