+++ src/tests/parties.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import PartiesStore, { Parties } from '../src/components/pages/party-page.jsx';
+import PartiesStore, { Parties } from '../components/pages/party-page.jsx';
 import {shallow, mount} from 'enzyme';
