+++ src/tests/signin.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 // import { SignIn } from '../src/components/pages/sign-in.jsx';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import SignInStore, { SignIn } from '../src/components/pages/sign-in.jsx';
+import SignInStore, { SignIn } from '../components/pages/sign-in.jsx';
 import {shallow, mount} from 'enzyme';
